Implemented Roadster Info request. Created Roadster Info model struct

+++ SpaceX Data/SpaceX Data/Models/Measurements/Measurement.swift
@@ -9,6 +9,6 @@
 import Foundation
 
 protocol Measurement {
-    var metric: Float { get }
-    var imperial: Float  { get }
+    var metric: Double { get }
+    var imperial: Double  { get }
 }
